Add `imagePort` to support Docker images not `EXPOSE`-ing port 80 (#383)
* Change mup.js.sample to use newer `docker` object-notation.

* Add `imagePort` to support Docker images not `EXPOSE`-ing port 80

Since newer Meteor Docker images are using non-root users to run the app (a commonly suggested security move in Node+Docker environments), it is no longer safe to assume that the Docker image will be exposing (nor may it be possible to EXPOSE) port 80.

This adds an `imagePort` setting in the Docker object (alongside the `image` setting) which defaults to `80` for compatibility with `kadirahq/meteord` (and `meteorhacks/meteord`) but adds the ability to change the *internal* port which is exposed by the docker image (i.e. the port that `--publish PORT:` will map to).