Replace npm-shrinkwrap with package-lock There once was a patch update to a dependency which broke mup if it was installed after that update. We started using npm-shrinkwrap to avoid that happening again.
New versions of NPM have started including dev and optional dependencies in npm-shrinkwrap, and always installing them when a package is installed globally. We had a script to fix the optional dependencies, but it will be more difficult to remove dev dependencies. For now we will switch back to using package-lock.json.